v9.1.1

+++ 52.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
-#Data alteração 21242
-#REV. 3
+#Data alteração 21244
+#REV. 4
 
+if [ -d "${MANAGER}/Kernel/" ] ; then
+echo -e "Validando ${MANAGER}/Kernel/"
 sudo find ${MANAGER}/Kernel/ -iname 'TH*.TR*' -exec rm -rf {} \; &
+fi
 
 echo -e "" &&
 echo -e "" &&
